[rename] styles 폴더 생성 - App.css 이동 - index.css 이동 - reset.css 이동
