+++ src/main.jsx
@@ -1,5 +1,5 @@
 import { createRoot } from "react-dom/client";
-import "./index.css";
+import "./styles/index.css";
 import App from "./App.jsx";
 import { BrowserRouter } from "react-router-dom";
